+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add begin/end addresses in ram search
 * Add byte array type in ram search
 * Add minimal hooks for xkbcommon to fix ruffle keyboard layout
+* Add a window to modify movie settings
 
 ### Changed